Split lib.api to multiple files
