+++ lib/api/core.py
@@ -1,15 +1,17 @@
 from distutils.version import LooseVersion as _version
 
-from rest_framework import exceptions, serializers
+from rest_framework import exceptions
 from rest_framework.renderers import JSONRenderer
 from rest_framework.versioning import AcceptHeaderVersioning, URLPathVersioning
 from rest_framework.negotiation import DefaultContentNegotiation
 from rest_framework.compat import unicode_http_header
 from rest_framework.utils import mediatypes
 
+
 # define this here as it's project dependent and not installation
 # there is no point to redefine it ever
 APLUS_JSON_TYPE = 'application/vnd.aplus+json'
+
 
 class _MediaType(mediatypes._MediaType):
     """
@@ -95,23 +97,3 @@ class APlusContentNegotiation(DefaultContentNegotiation):
                 accept = str(mt)
             accepts.append(accept)
         return accepts
-
-
-class ListSerializerMixin(object):
-    def get_serializer_class(self):
-        if self.action == 'list':
-            return getattr(self, 'listserializer_class', self.serializer_class)
-        return super(ListSerializerMixin, self).get_serializer_class()
-
-
-class HtmlViewField(serializers.ReadOnlyField):
-    def __init__(self, *args, **kwargs):
-        super(HtmlViewField, self).__init__(*args, **kwargs)
-
-    def get_attribute(self, obj):
-        return obj
-
-    def to_representation(self, obj):
-        request = self.context['request']
-        url = obj.get_absolute_url()
-        return request.build_absolute_uri(url)
