+++ lib/api/mixins.py
@@ -1,0 +1,5 @@
+class ListSerializerMixin(object):
+    def get_serializer_class(self):
+        if self.action == 'list':
+            return getattr(self, 'listserializer_class', self.serializer_class)
+        return super(ListSerializerMixin, self).get_serializer_class()
